refactor: update starter code to use vite react boilerplate
